There are 3 different versions of cricket, however we're going to focus on (T20) 20 over cricket.
Issues i'm noticing, after selecting a 'Home School' there are no options in the 'Home Team' drop down. The 'Away Team' is grey out and so is the date.

Match Type should be the first option.

'Age groups' should have some sort of association and link with the 'Home Team'.

Location should be based on 'Field' options.

It's all a bit of a mess and very untidy. Make the UI and UX more dynamic